// src/types/types.ts
export enum TYPES {
    NULL = '[object Null]',
    STRING = '[object String]',
    UNDEFINED = '[object Undefined]',
    NUMBER= '[object Number]',
    OBJECT= '[object Object]',
    ARRAY = '[object Array]',
    FUNCTION= '[object Function]',
    DATE ='[object Date]',
    BOOLEAN = '[object Boolean]',
    ARGUMENTS = '[object Arguments]',
    MAP = '[object Map]',
    WEAKMAP = '[object WeakMap]',
    WEAKSET = '[object WeakSet]',
    SET= '[object Set]',
    SYMBOL ='[object Symbol]',
    REG = '[object RegExp]'
}